+++ ExilenceNextApp/src/store/domains/account.ts
@@ -5,7 +5,6 @@ import { fromStream } from 'mobx-utils';
 import { of, Subject, throwError, timer } from 'rxjs';
 import { catchError, map, mergeMap, retryWhen, switchMap, takeUntil } from 'rxjs/operators';
 import uuid from 'uuid';
-
 import { IAccount } from '../../interfaces/account.interface';
 import { IApiAccount } from '../../interfaces/api/api-account.interface';
 import { IApiProfile } from '../../interfaces/api/api-profile.interface';
@@ -86,12 +85,15 @@ export class Account implements IAccount {
   }
 
   @action
-  queueSnapshot() {
+  queueSnapshot(milliseconds?: number) {
     fromStream(
-      timer(rootStore.settingStore.autoSnapshotInterval).pipe(
+      timer(milliseconds ? milliseconds : rootStore.settingStore.autoSnapshotInterval).pipe(
         map(() => {
           if (this.activeProfile && this.activeProfile.readyToSnapshot) {
             this.activeProfile.snapshot();
+          } else {
+            this.dequeueSnapshot();
+            this.queueSnapshot(10 * 1000);
           }
         }),
         takeUntil(this.cancelled)